Use arbitrary precision representations until we decide how to handle precision.
git-svn-id: https://svn.sable.mcgill.ca/mclab/Project@157 89b902b4-d83f-0410-b942-89f7a419ffdb

Migrated from Sable/mclab@0c2ff4223a2c9764c1509c8f484ad976183c8371

+++ languages/Natlab/src/natlab/IntNumericLiteralValue.java
@@ -1,18 +1,20 @@
 package natlab;
 
+import java.math.BigInteger;
+
 /**
  * The value of a numeric literal that is definitely not a FP number.
  */
 public abstract class IntNumericLiteralValue extends NumericLiteralValue {
-	private final Integer value;
+	private final BigInteger value;
 
-	public IntNumericLiteralValue(String text, Integer value, boolean isImaginary) {
+	public IntNumericLiteralValue(String text, BigInteger value, boolean isImaginary) {
 		super(text, isImaginary);
 		this.value = value;
 	}
 
 	@Override
-	public Integer getValue() {
+	public BigInteger getValue() {
 		return value;
 	}
 }
